Move jest tests to their own file and ignore test util folder in codeclimate
