+++ .codeclimate.json
@@ -4,6 +4,7 @@
     "**/*.test.ts",
     "extension/src/test/fixtures",
     "extension/src/test/suite",
+    "extension/src/test/util",
     "extension/src/vscode.proposed.d.ts",
     "webview/src/react-table-config.d.ts"
   ]
