+++ demo/params.yaml
@@ -1,3 +1,7 @@
 seed: 473987
 lr: 0.0005
 weight_decay: 0
+nested:
+  nested:
+    nested:
+      test: true
