+++ extension/src/cli/reader.test.ts
@@ -5,7 +5,7 @@ import { Disposable, Disposer } from '@hediet/std/disposable'
 import { CliResult, CliStarted } from '.'
 import { CliReader } from './reader'
 import { createProcess } from '../processExecution'
-import { getFailingMockedProcess, getMockedProcess } from '../test/util'
+import { getFailingMockedProcess, getMockedProcess } from '../test/util/jest'
 import { getProcessEnv } from '../env'
 import complexExperimentsOutput from '../test/fixtures/complex-output-example'
 import { Config } from '../config'
